fix: validations ok

+++ MobileChallenge/Models/Items.swift
@@ -3,11 +3,11 @@ import Foundation
 class Items: Codable {
     
     let name: String
-    let value: Double
+    let value: Int
     let amount: Int
     var total: Double?
     
-    init(name: String, value: Double, amount: Int){
+    init(name: String, value: Int, amount: Int){
         self.name = name
         self.value = value
         self.amount = amount
